+++ src/common/devtools/combineDevtools.ts
@@ -4,8 +4,8 @@ export function combineDevtools(...devtools: DevtoolsLike[]): DevtoolsLike {
     return {
         state(name, initState) {
             const updaters = devtools.map((d) => d.state(name, initState));
-            return (newState) => {
-                updaters.forEach((update) => update(newState));
+            return (newState, actionName) => {
+                updaters.forEach((update) => update(newState, actionName));
             };
         },
     };