+++ src/common/devtools/types.ts
@@ -1,5 +1,5 @@
 export interface DevtoolsStateLike<T = any> {
-    (newState: T): void;
+    (newState: T, actionName?: string): void;
 }
 export interface DevtoolsLike {
     state<T>(name: string, initState: T): DevtoolsStateLike<T>;