+++ src/signals/types/SignalOptions.ts
@@ -1,10 +1,10 @@
 export interface SignalLifecycleHook<T = any> {
     onInit?: (value: T) => void;
-    onChange?: (newValue: T) => void;
+    onChange?: (newValue: T, actionName?: string) => void;
     onDispose?: () => void;
 }
 
-export type TBeforeDevtoolsPushFn<T = any> = (newValue: T, push: (v: T) => void) => void;
+export type TBeforeDevtoolsPushFn<T = any> = (newValue: T, push: (v: T) => void, actionName?: string) => void;
 
 export interface SignalOptions<T = any> {
     key?: string;