+++ src/signals/types/signals.types.ts
@@ -11,8 +11,8 @@ export interface ReadableSignalFnLike<T> extends ReadableSignalLike<T> {
 }
 
 export interface WriteableSignalLike<T> {
-    set(value: T): void;
-    update(updater: (value: T) => T): void;
+    set(value: T, actionName?: string): void;
+    update(updater: (value: T) => T, actionName?: string): void;
 }
 
 export interface ClearableSignalLike<_T> {